Pessoa 100% e Gravador de Imagens 100%

+++ src/br/com/crm/model/service/AreaInteresseService.java
@@ -84,7 +84,7 @@ public class AreaInteresseService {
 	 * @param id
 	 * @return
 	 */
-	public AreaInteresse buscarAreaInteressePorId(Long id) {
+	public AreaInteresse buscarAreaInteressePorId(Integer id) {
 		return manager.find(AreaInteresse.class, id);
 	}
